Santa, Elf, Reindeer and groups. Currently the "leave and do something else" part of santa problem is not correctly modelled IMO.
This is a solution where Elves and Reindeer
and modelled as joining a group. The groups
and santa a cowns that are waited upon
for access to either join or deliver toys
or meet in the study. After the activity,
the members involved (excluding Santa) go
off and do something for a while.